% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_abiotic.R, R/fig_biotic.R
\name{fig_biotic}
\alias{fig_biotic}
\title{Make figures for biotic data}
\usage{
fig_biotic(lang = "fr")

fig_biotic(lang = "fr")
}
\description{
Make figures for biotic data

Make figures for biotic data
}
